## Description
Release the stable release 'HereBeMonsters' (aka the glorious and infamous version `H`). This release adds:

### Major changes
- Add Plagiarism checkers (#486) (#513) (#536) (#555) (#508) (#556) (#645) (#576)
  It is now possible to check for plagiarism in CodeGrade. This enables privacy aware plagiarism checking. It is possible to use check against old CodeGrade assignment and upload base code and old submissions that are not in CodeGrade. For more information see our documentation.

- Make it possible give grades higher than ten (#500)
  Teachers can now indicate that students can receive a grader higher than 10 for an assignment, making it possible to create assignments with bonus points in CodeGrade. When using within LTI this requires a new LTI parameter.
  
  You should add the following to the `<blti:custom>` section of your canvas LTI config for CodeGrade:

  ```xml
  <lticm:property name="custom_canvas_points_possible">
    $Canvas.assignment.pointsPossible
  </lticm:property>
  ```
  

### Minor changes
- Change homepage to login screen (#480) (#496) (#501) (#557) (#616)
  The homepage has been improved to show all your courses and assignments at a glance when logged in.
- Allow .tar.xz archives to be uploaded (#492)
  This further improves the flexibility CodeGrade gives students when handing in submissions.
- Fix infinite loop overview mode (#477)
  In some combinations of permissions loading the overview mode resulted in an infinite loader.
- Add general feedback tab to overview mode (#523)
  This further decreases the chance that students will miss any of their feedback.
- Fix broken matchFiles function (#528) (#550)
  This fixes a bug that files changed inside a directory would not show up in the overview mode.
- Improve speed of diffing by using another library (#529)
  Viewer the diff between two large files is a lot faster!
- Remove the option to automatically generate keys (#554)
  It is no longer possible to generate the `secret_key` or `lti_secret_key` configuration options. Please update your config accordingly.
- Rewrite snippets manager (#551)
  This rewrite should make creating, using, deleting and updating snippets faster and more reliable.
- Drastically improve the experience of CodeGrade on mobile (#558)
  It is now way easier to use CodeGrade on mobile.
- Filter users in the user selector (#553)
  When selecting users (when uploading for others, or adding to courses) only show users will be shown that can be selected.
- Improve handling of LTI (#561) (#588)
  A complete rewrite of LTI backend handling. This should improve the stability of passbacks by a lot. This also guarantees that the submission date in Canvas and CodeGrade will match exactly.
  
  This also adds a new convenience route `/api/v1/lti/?lms=Canvas` to get lti config for the given LMS (Canvas only supported at the moment).
- Add items to the sidebar conditionally (#578) (#580) (#600)
  Depending on what page you are you will get extra items in the sidebar to help quick navigation. Currently plagiarism cases and submissions are added depending on the page.
- Start caching submissions (#643) (#636)
  Submissions are cached in the front-end so changing between the codeviewer and submissions list is now way quicker.
- Ensure all rubric rows have a maximum amount of >= 0 points (#579)
  It is no longer allowed to have rows in a rubric where the maximum possible score is < 0. If you needed this to create rubrics with bonus categories simply use the 'Max points' option in the rubric editor. All existing rubrics are not changed.


### Small changes and bugfixes
- Use new logos (#481) (#506)
  This updates our logo to the newest and greatest version!
- Various small bugs in the sidebar
- Add a minimum duration on the permission manager loaders (#521)
  This makes it clearer that permissions are actually updated.
- Throw an API error when a rubric row contains an empty header (#535)
  This is a backwards incompatible API change, however it doesn't change anything for the frontend.
- Fix horizontal overflow on codeviewer (#518)
  The codeviewer would sometimes overflow creating a vertical scrollbar when displaying files containing a large amount of consecutive tabs.
- Check if an assignment is loaded before getting its course (#549)
  In some rare cases LTI launches would fail be cause assignments were not loaded correctly.
- Add structured logging setup (#546)
  This makes it easier to follow requests and debug issues.
- Fix general feedback line wrapping (#570)
  Giving long lines as general feedback should be displayed correctly to the user now.
- Add manage assignment button to submission list (#574)
  It is now possible to easily navigate to the manage assignment page from the submissions list.
- Start using enum to store permissions in the backend (#571)
  Most routes will be faster by this design change.
- Improve filetree design (#599) (#611)  (#587)
  It is now easier to spot additions, changes and deletion directly in the filetree.
- Add ``<noscript>`` tag (#613)
  An error message will be displayed when javascript is disabled.
- Improve speed of filetree operations (#623)
  Loading large filetrees is now way quicker by using smarter data-structures.
- Add health route (#593)
  It is now possible to more easily monitor the health of your CodeGrade instance.
- Fix fontSize & contextAmount on submission page (#633)
  Sometimes the fields would show up empty, this shouldn't happen anymore!
- Replace submitted symlinks with actual files (#627)
  When a student uploads an archive with symlinks the student is warned and all symlinks are replaced by files explaining that the original files were symlinks but that those are not supported by CodeGrade.
- Fix grade history popover boundary (#625)
  The grade history would sometimes show up outside the screen, but no more!
- Make it impossible to submit empty archives (#622)
  A error is shown when a student tries to submit an archive without files.
- Show toast when local-storage doesn't work (#607)
  When a user has no local-storage available a warning is shown so the user knows that their experience might be sub-optimal.
- Show author of general feedback and line comments (#564) (#605)
  The author of all general feedback and line comments is displayed to the user. Only users with the `can_see_assignee` permission will see authors.
- Justify description popover text (#596)
  The text in descriptions is now justified and their popups will only show when the 'i' is clicked.
- Only submit rubric items or normal grade (#589)
  In some rare cases overriding rubrics would result in a race condition, resulting in wrong case.
- Redesign the download popover on the submission page (#595)
  This new design looks way better, but you tell us!
- Only show overview mode when you have permission to see feedback (#563)
  When you don't have permission to see feedback the overview mode will never be shown.
- Various other performance improvements (#566)
  We always strive for the best performance possible, and again in this release we increased the performance of CodeGrade! :tada:
- Make sure codeviewer is full width on medium pages (#591)
  This makes it easier to review and display code on smaller screens.
- Use custom font in toasted actions (#614)
  It is now always possible to close toasts, even when your font cannot display '✖'

+++ psef/v1/code.py
@@ -2,7 +2,7 @@
 This module defines all API routes with the main directory "code". Thus the
 APIs are used to manipulate student submitted code and the related feedback.
 
-:license: AGPLv3, see LICENSE for details.
+SPDX-License-Identifier: AGPL-3.0-only
 """
 
 import shutil
@@ -26,6 +26,7 @@ from psef.helpers import (
 )
 
 from . import api
+from ..permissions import CoursePermission as CPerm
 
 _HumanFeedback = models.Comment
 _LinterFeedback = t.MutableSequence[t.Tuple[str, models.LinterComment]]  # pylint: disable=invalid-name
@@ -49,9 +50,10 @@ def put_comment(code_id: int, line: int) -> EmptyResponse:
     :raises PermissionException: If the user can not can grade work in the
                                  attached course. (INCORRECT_PERMISSION)
     """
-    comment = db.session.query(models.Comment).filter(
-        models.Comment.file_id == code_id, models.Comment.line == line
-    ).one_or_none()
+    comment = db.session.query(
+        models.Comment
+    ).filter(models.Comment.file_id == code_id,
+             models.Comment.line == line).one_or_none()
 
     def __get_comment() -> str:
         content = ensure_json_dict(request.get_json())
@@ -60,14 +62,14 @@ def put_comment(code_id: int, line: int) -> EmptyResponse:
 
     if comment:
         auth.ensure_permission(
-            'can_grade_work', comment.file.work.assignment.course_id
+            CPerm.can_grade_work, comment.file.work.assignment.course_id
         )
 
         comment.comment = __get_comment()
     else:
         file = helpers.get_or_404(models.File, code_id)
         auth.ensure_permission(
-            'can_grade_work',
+            CPerm.can_grade_work,
             file.work.assignment.course_id,
         )
 
@@ -108,7 +110,7 @@ def remove_comment(code_id: int, line: int) -> EmptyResponse:
     )
 
     auth.ensure_permission(
-        'can_grade_work', comment.file.work.assignment.course_id
+        CPerm.can_grade_work, comment.file.work.assignment.course_id
     )
     db.session.delete(comment)
     db.session.commit()
@@ -119,8 +121,7 @@ def remove_comment(code_id: int, line: int) -> EmptyResponse:
 @api.route("/code/<int:file_id>", methods=['GET'])
 @auth.login_required
 def get_code(file_id: int) -> t.Union[werkzeug.wrappers.Response, JSONResponse[
-    t.Union[t.Mapping[str, str], models.File, _FeedbackMapping]
-]]:
+    t.Union[t.Mapping[str, str], models.File, _FeedbackMapping]]]:
     """Get data from the :class:`.models.File` with the given id.
 
     .. :quickref: Code; Get code or its metadata.
@@ -157,7 +158,7 @@ def get_code(file_id: int) -> t.Union[werkzeug.wrappers.Response, JSONResponse[
         return jsonify(file)
     elif get_type == 'feedback':
         return jsonify(get_feedback(file, linter=False))
-    elif get_type == 'pdf' or get_type == 'file-url':
+    elif get_type in ('pdf', 'file-url'):
         return jsonify({'name': get_file_url(file)})
     elif get_type == 'linter-feedback':
         return jsonify(get_feedback(file, linter=True))
@@ -291,6 +292,27 @@ def delete_code(file_id: int) -> EmptyResponse:
                 APICodes.INVALID_STATE,
                 400,
             )
+        if db.session.query(
+            models.PlagiarismMatch.query.filter(
+                (models.PlagiarismMatch.file1_id == code.id)
+                | (models.PlagiarismMatch.file2_id == code.id)
+            ).exists()
+        ).scalar():
+            # TODO: This leaks information. The question is if this is really
+            # a big issue. To stop leaking information all the other error
+            # messages also need to be adjusted, and even then: the other
+            # properties can probably be checked (if it has comments of
+            # children) so we would still leak some information. Another better
+            # option would be to only mark the file as deleted.
+            raise APIException(
+                (
+                    'You cannot delete this code as it is implicated in'
+                    ' plagiarism'
+                ),
+                f'The file "{file_id}" is implicated in a plagiarism match.',
+                APICodes.INVALID_STATE, 400
+            )
+
         code.delete_from_disk()
         db.session.delete(code)
     elif code.fileowner == models.FileOwner.both:
@@ -323,20 +345,24 @@ def split_code(
     old_id = code.id
     old_diskname = None if code.is_directory else code.get_diskname()
     db.session.flush()
-    code = db.session.query(models.File).get(code.id)
+    code = t.cast(models.File, db.session.query(models.File).get(code.id))
+    assert code is not None
 
     db.session.expunge(code)
     make_transient(code)
-    code.id = None
+    code.id = None  # type: ignore
     db.session.add(code)
     db.session.flush()
 
     code.fileowner = new_owner
     if not code.is_directory:
+        assert old_diskname is not None
         _, code.filename = psef.files.random_file_path()
         shutil.copyfile(old_diskname, code.get_diskname())
     else:
-        redistribute_directory(code, models.File.query.get(old_id))
+        redistribute_directory(
+            code, t.cast(models.File, models.File.query.get(old_id))
+        )
 
     return code
